Add a pre-commit hook to use the Ruff linter
Ruff is a tool that is able to replace multiple other Python focused
linters/formatters. We are only interested in the linter portion of
its functionality right now. The initial configuration file we are
using duplicates the one we currently use for flake8. Since Ruff can
act as a drop-in replacement for flake8 this should give us
consistent behavior around linting results for our codebases.